requirements.txt
google-cloud-aiplatform
flask
openai 
python-dotenv
pinecone-client
sentence-transformers
